[#S11P31D108-300]: QA 반영, 타임캡슐 리스트 api (근데 엉터리임 다시 해야함)

See merge request s11-final/S11P31D108!46

+++ android/presentation/src/main/java/com/familring/presentation/screen/timecapsule/TimeCapsuleScreen.kt
@@ -68,7 +68,7 @@ fun TimeCapsuleScreen(
     state: TimeCapsuleUiState,
     getTimeCapsuleStatus: () -> Unit = {},
     createTimeCapsuleAnswer: (String) -> Unit = {},
-    getTimeCapsules: () -> Unit = {},
+    getTimeCapsules: (Int) -> Unit = {},
     onShowSnackBar: (message: String) -> Unit = {},
 ) {
     val tabs = listOf("작성", "목록")
